+++ app/Filament/Resources/CiQualityGateResource/Pages/CreateCiQualityGate.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\CiQualityGateResource\Pages;
+
+use App\Filament\Resources\CiQualityGateResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateCiQualityGate extends CreateRecord
+{
+    protected static string $resource = CiQualityGateResource::class;
+}