+++ app/Filament/Resources/CiQualityGateResource/Pages/ViewCiQualityGate.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\CiQualityGateResource\Pages;
+
+use App\Filament\Resources\CiQualityGateResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewCiQualityGate extends ViewRecord
+{
+    protected static string $resource = CiQualityGateResource::class;
+}